don't reference dart:io in code compiling for the web
